src/requirements.txt
azure-identity~=1.9.0
python-dateutil~=2.8.2
fastapi~=0.75.2
fastapi-health~=0.4.0
fastapi-utils~=0.2.1
json-logging~=1.3.0
msgraph-core~=0.2.2
pydantic~=1.9.0
starlette-exporter~=0.12.0
uvicorn~=0.17.1
